requirements.txt
botbuilder-core>=4.12.0
asyncio>=3.4.3
aiohttp>=3.6.2
cookiecutter>=1.7.0
botbuilder-integration-aiohttp>=4.12.0
botbuilder-dialogs>=4.11.0
botbuilder-ai>=4.11.0
